# API/AdminPanelApi/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api",
    "version": "v1"
  },
  "paths": {
...
    "/api/Section": {
      "get": {
        "tags": [
          "Section"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Section"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "ContentType",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "ContentDisposition",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Headers",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Length",
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "formData",
            "name": "Name",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "FileName",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "SectionName",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Index",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Text",
            "type": "array",
            "items": {
              "$ref": "#/definitions/Text"
            }
          },
          {
            "in": "formData",
            "name": "Image.ImagePath",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Image.SectionId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Image.Section.Id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Image.Section.SectionName",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Image.Section.Index",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Image.Section.Text",
            "type": "array",
            "items": {
              "$ref": "#/definitions/Text"
            }
          },
          {
            "in": "formData",
            "name": "Image.Section.Image.ImagePath",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Image.Section.Image.SectionId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Image.Section.Image.Section"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Section"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "ContentType",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "ContentDisposition",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Headers",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Length",
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "formData",
            "name": "Name",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "FileName",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "SectionName",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Index",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Text",
            "type": "array",
            "items": {
              "$ref": "#/definitions/Text"
            }
          },
          {
            "in": "formData",
            "name": "Image.ImagePath",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Image.SectionId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Image.Section.Id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Image.Section.SectionName",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Image.Section.Index",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Image.Section.Text",
            "type": "array",
            "items": {
              "$ref": "#/definitions/Text"
            }
          },
          {
            "in": "formData",
            "name": "Image.Section.Image.ImagePath",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Image.Section.Image.SectionId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Image.Section.Image.Section"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Section/{id}": {
      "delete": {
        "tags": [
          "Section"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Steps": {
      "get": {
        "tags": [
          "Steps"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Steps"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Step"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Steps"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Step"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Steps/{id}": {
      "delete": {
        "tags": [
          "Steps"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
...
    "SectionImages": {
      "type": "object",
      "properties": {
        "imagePath": {
          "type": "string"
        },
        "sectionId": {
          "format": "int32",
          "type": "integer"
        },
        "section": {
          "$ref": "#/definitions/Section"
        }
      }
    },
    "Section": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "sectionName": {
          "type": "string"
        },
        "index": {
          "format": "int32",
          "type": "integer"
        },
        "text": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Text"
          }
        },
        "image": {
          "$ref": "#/definitions/SectionImages"
        }
      }
    },
    "Text": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "content": {
          "type": "string"
        },
        "sectionId": {
          "format": "int32",
          "type": "integer"
        },
        "section": {
          "$ref": "#/definitions/Section"
        }
      }
    },
    "Step": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "titleEn": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "descriptionEn": {
          "type": "string"
        },
        "active": {
          "type": "boolean"
        }
      }
    }
  }
}